move: separate collsion resolver classes into files
